+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Steptionary changelog
 
+### [v0.0.2](https://github.com/goriol/Steptionary/tree/0.0.2)
+
+* Setup cache when a window is activated
+* Suggest completions based on cached steps
+
 ### [v0.0.1](https://github.com/goriol/Steptionary/tree/0.0.1)
 
 * Refresh cache when a feature is saved